feat: edited driver_test json file and change an existing function
