+++ src/App.js
@@ -1,6 +1,6 @@
 import SelectExam from "./components/SelectExam";
 import LabExam from "./components/LabExam";
-import Test from "./test/html_parse_img/test"
+import ParsingImageTest from "./test/html_parse_img/parsingImageTest"
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 function App() {
   return (
@@ -10,7 +10,7 @@ function App() {
         <Routes>
           <Route path="/" element={<SelectExam />} />
           <Route path="/lab" element={<LabExam />} />
-          <Route path="/html_parse" element={<Test />} />
+          <Route path="/htmlParse" element={<ParsingImageTest />} />
         </Routes>
       </div>
     </BrowserRouter>
