+++ src/components/multiImage.js
@@ -1,0 +1,6 @@
+<div style={{ width: "1000px" }}>
+  <img src="https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2Fnspuc%2FbtsFt5GRCru%2FQyCoV4BBeEMoBGyNWoMCRk%2Fimg.png"></img>
+  <img src="https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2Fnspuc%2FbtsFt5GRCru%2FQyCoV4BBeEMoBGyNWoMCRk%2Fimg.png"></img>
+  <img src="https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2Fnspuc%2FbtsFt5GRCru%2FQyCoV4BBeEMoBGyNWoMCRk%2Fimg.png"></img>
+  <img src="https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2Fnspuc%2FbtsFt5GRCru%2FQyCoV4BBeEMoBGyNWoMCRk%2Fimg.png"></img>
+</div>
